Merged PR 603: Publish secrets through the key vault flex volume
Creates a KeyVault to hold the secrets for the delivery service, and exposes them using the KeyVault flex volume using AAD Pod identity. These secrets exist SxS with the k8s secrets, and the app does not consume them yet.

Updated the deployment instructions with all the necessary details. Also adds yaml files for setting up pod identity and updates the delivery yaml to use pod identity, mount the volumes and sets the fsGroup to allow the delivery user to access the mounted volumes.

The change to the service's port is due to an issue encountered recently where the app wouldn't be able to bind port 80. This is not caused by the changes, as I've rolled them back and it would still fail. Not sure what this is about but will need to be investigated.

Related work items: #8117, #8148, #8150