Merged PR 645: Address e2e testing issues
tweak the serialization and deserialization used by workflow to communicate with the package service to match the implicit casing requirements
remove the "clusterIP: none" from the service specs. this is supposed to be used for headless services and these arent https://kubernetes.io/docs/concepts/services-networking/service/#headless-services. communication worked before because there were no port mappings, but with port mappings this makes connections through the service to fail.