add ca warning note for basic + catchup minor fixes in master (#161)
* fixing the testing section so that it works (#154)

the previous approach assumed that you could set the delivery_id, which you can't. As a result, the confirmation listing call fails with a 404. this captures the delivery_id and then uses that value to see the others. Fixes -- or addresses -- #132

* required permissions note (#155)

add a note to the deployment steps indicating required permissions

* include CA message warning users about using self-signed certificates

closes: #190580

* change wording for role assignments

Co-authored-by: Page Writer <18742226+PageWriter-MSFT@users.noreply.github.com>

* align with feedback from master

Co-authored-by: Ralph Squillace <rasquill@microsoft.com>
Co-authored-by: Page Writer <18742226+PageWriter-MSFT@users.noreply.github.com>