Retire Fluentd-ES integration (#194)
closes: #330826 #131

we made the decision of retiring this integration with fluentd-es because for the moment since this is currently out of date. We keep encouraging the users to use OOS solutions, and in particular Fluentd-ES from [our docs](https://docs.microsoft.com/en-us/azure/architecture/microservices/logging-monitoring).

Additionally it helps us to streamline a bit the instructions as well as save our users from importing these images to private registry because of the reliance on public container images.